refactor(spec): split spectest helper files
