+++ crates/spectest/src/test_case/writer.rs
@@ -1,0 +1,3 @@
+pub fn write<'wasm>(modules: Vec<Module<'wasm>>) {
+    
+}
